requirements.txt
flask==3.0.0
openai>=1.50.0
reportlab==4.0.9
python-dotenv==1.0.0
PyPDF2==3.0.1
werkzeug==3.0.1
python-docx==1.1.0